+++ server/middlewares/remote-authenticate.js
@@ -97,7 +97,7 @@
         logEntry + '\n',
         {
           encoding: 'utf8',
-          mode: 0o644,
+          mode: 0o600,
           flag: 'a'
         },
         function (err) {