+++ server/middlewares/user-authenticate.js
@@ -84,7 +84,7 @@
         logEntry + '\n',
         {
           encoding: 'utf8',
-          mode: 0o644,
+          mode: 0o600,
           flag: 'a'
         },
         function (err) {